MiniWin 1.0-pre-11
Real 32-bit truecolor: 640x480, genuine direct-color VBE mode, any
24-bit RGB value -- not a 256-entry palette anymore.

Split the bootloader into two stages (boot/boot.asm, exactly 512 bytes;
boot/stage2.asm, everything else) because finding a real truecolor mode
properly -- walking the BIOS's own VBE mode list (INT10h AX=4F00h for
the Controller Info Block, then AX=4F01h per candidate mode) checking
each one's actual XResolution/YResolution/BitsPerPixel/MemoryModel
against what's wanted, rather than just requesting a fixed, hoped-for
mode number the way the old 640x400x8bpp setup did -- stopped fitting
in a single boot sector's budget. There's no equivalent to mode 13h's
(or even VBE's own 0100h's) stable mode-number convention at truecolor
depths; different cards and emulators assign that mode differently, so
enumeration is the only portable way to find it. A machine offering no
matching mode fails cleanly with a clear message rather than silently
falling back to a second, permanently-maintained 8bpp rendering
pipeline.

kernel/vga.h rewritten for direct color: backbuf is u32 (a packed RGB
pixel) instead of u8 (a palette index), built via a new RGB(r,g,b)
helper that honors whatever red/green/blue bit-field positions the
chosen VBE mode actually reports (read from the Mode Info Block,
verified against the universal 16/8/0 layout at boot rather than
assumed). vga_set_standard_palette()/col_cube()/col_gray() -- the old
DAC-palette-programming code -- removed outright: there's no palette to
program in direct-color mode, and carrying it as dead code serves
nobody. COL_* constants changed from palette indices to real 24-bit RGB
values (the same classic 16-color EGA/VGA RGB numbers, just expressed
directly) -- which meant the 150+ call sites across kernel.c that
reference COL_BLACK, COL_WHITE, and so on didn't need to change at all,
only the constant definitions and the draw primitives' parameter types
(u8 -> u32) did.

The 1.2MB truecolor backbuffer (640*480*4 bytes) doesn't fit in this
kernel's usual sub-640KB footprint -- placed instead at a fixed
physical address (4MB) using the same 'paging is off, a raw address is
just a pointer' trick the framebuffer pointer itself already relied on.
Unlike blindly trusting that address is safe RAM, stage2.asm now probes
actual memory size via INT15h AX=E801h and the kernel checks it
(vga_verify_memory_safe()) before ever writing through the backbuffer
pointer, refusing to boot with a clear serial message instead of
risking silent corruption on a machine too small for it -- the same
'verify the assumption instead of trusting it' instinct that caught the
BSS/stack collision and GCM's undersized scratch buffer earlier this
project's own history.

Verified pixel-perfect: screenshotted a live QEMU boot and read back
exact RGB values for known UI colors (LCYAN background, LGRAY taskbar)
matching their COL_* definitions exactly, then filled a rectangle with
an arbitrary off-palette color (no relation to the classic 16) and
confirmed it rendered back byte-for-byte -- real 24-bit color depth,
not a wider-looking reinterpretation of the same 16 colors. GUI
rendering (windows, title bars, Hangul text, bevel buttons) confirmed
unaffected by the cutover via a second live screenshot of SETTING.EXE
open.

BSS headroom actually improved despite the far larger nominal
backbuffer (250KB static array -> 1.2MB fixed-address pointer): moving
the buffer out of .bss entirely freed far more space than the four-byte
pointer that replaced it costs, taking build.sh's own BSS/stack margin
check from 13KB of headroom to 268KB.

+++ kernel/font.h
@@ -108,7 +108,7 @@ static inline char to_upper(char c) {
     return c;
 }
 
-static inline void font_draw_char(int x, int y, char c, u8 color) {
+static inline void font_draw_char(int x, int y, char c, u32 color) {
     const u8 *glyph;
 
     if (c >= 'a' && c <= 'z') {
@@ -129,7 +129,7 @@ static inline void font_draw_char(int x, int y, char c, u8 color) {
     }
 }
 
-static inline void font_draw_string(int x, int y, const char *s, u8 color) {
+static inline void font_draw_string(int x, int y, const char *s, u32 color) {
     int cx = x;
     while (*s) {
         if (*s == '\n') { cx = x; y += 9; s++; continue; }
@@ -147,7 +147,7 @@ static inline void font_draw_string(int x, int y, const char *s, u8 color) {
  * font just so a handful of letters can stand up straight.
  * (x,y) is the top-left corner of the resulting 8x8 box, same as
  * font_draw_char -- it just happens to contain a sideways letter. */
-static inline void font_draw_char_vertical(int x, int y, char c, u8 color) {
+static inline void font_draw_char_vertical(int x, int y, char c, u32 color) {
     char u = to_upper(c);
     if (u < 0x20 || u > 0x5A) return; /* lowercase-only glyphs don't get a rotated cousin; not worth it */
     const u8 *glyph = font8x8[(int)u - 0x20];
@@ -166,7 +166,7 @@ static inline void font_draw_char_vertical(int x, int y, char c, u8 color) {
  * reads bottom-to-top -- perfect for a thin vertical strip along the
  * left edge of a popup menu, or for silently judging anyone still
  * running this OS at 320x200. */
-static inline void font_draw_string_vertical(int x, int y_bottom, const char *s, u8 color) {
+static inline void font_draw_string_vertical(int x, int y_bottom, const char *s, u32 color) {
     int cy = y_bottom;
     while (*s) {
         font_draw_char_vertical(x, cy, *s, color);
